refactor(cliproxy): extract gh orchestration into setup/gh.ts
Move CommandResult, withSpinner, runGh, isGhRateLimitError, withGhRetry,
assertGhInstalled, assertGhAuthenticated, assertRepoAccess, listExistingGhNames,
createManagementApiKey, deleteManagementApiKey, and applyGhValue into a focused
submodule. The assertion helpers become exported so the orchestrator (Unit 9)
can dependency-inject them via RunSetupDeps.

setup.ts shrinks from 1114 to 884 LOC. Adds a TDD-first applyGhValue test that
locks in the PR #102 stdin-pipe-not-body invariant (715 total, +1). Compensating
rollback regression tests deferred to Unit 9 where the runSetupCommand boundary
they target is created.
