+++ packages/cli/src/commands/cliproxy/setup.test.ts
@@ -9,13 +9,11 @@ import {
   formatDryRunPreview,
   formatWorkflowSnippet,
   interpretGhContentResult,
-  isGhRateLimitError,
   mustConfirmDestructive,
   registerCliproxySetup,
   runSmokeTest,
   validateSetupOptions,
   verifyModelsAvailable,
-  withGhRetry,
 } from './setup'
 import {getHarnessTemplate} from './setup/templates'
 
@@ -303,62 +301,6 @@ describe('cliproxy setup helpers', () => {
       ].join('\n')
       /* eslint-enable no-template-curly-in-string */
       expect(snippet).toBe(expected)
-    })
-  })
-
-  describe('isGhRateLimitError', () => {
-    it('returns true when text contains "rate limit"', () => {
-      expect(isGhRateLimitError('API rate limit exceeded')).toBe(true)
-    })
-
-    it('is case-insensitive', () => {
-      expect(isGhRateLimitError('You have exceeded a secondary RATE LIMIT')).toBe(true)
-    })
-
-    it('returns false for unrelated error messages', () => {
-      expect(isGhRateLimitError('Not Found (HTTP 404)')).toBe(false)
-    })
-
-    it('returns false for an empty string', () => {
-      expect(isGhRateLimitError('')).toBe(false)
-    })
-
-    it('returns false for a connection timeout', () => {
-      expect(isGhRateLimitError('connection timeout')).toBe(false)
-    })
-  })
-
-  describe('withGhRetry', () => {
-    it('returns the value when fn succeeds immediately', async () => {
-      const result = await withGhRetry('test label', async () => 'ok', false)
-
-      expect(result).toBe('ok')
-    })
-
-    it('re-throws non-rate-limit errors without querying the reset time', async () => {
-      const queryReset = async (): Promise<string> => {
-        throw new Error('queryReset should not have been called')
-      }
-      const err = new Error('some other error')
-
-      await expect(withGhRetry('test label', async () => Promise.reject(err), false, queryReset)).rejects.toThrow(
-        'some other error',
-      )
-    })
-
-    it('re-throws with reset time appended in non-interactive mode on rate limit', async () => {
-      const queryReset = async (): Promise<string> => '2:30 PM'
-
-      await expect(
-        withGhRetry(
-          'test label',
-          async () => {
-            throw new Error('API rate limit exceeded for url')
-          },
-          false,
-          queryReset,
-        ),
-      ).rejects.toThrow('resets at 2:30 PM')
     })
   })
